pyproject.toml
[tool.poetry]
name = "kafka-desktop-consumer"
version = "1.1"
description = "A desktop consumer for kafka topics, that supports embedded, schema registry and more."
authors = ["Nimja"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
avro-python3 = "^1.10.2"
confluent-kafka = "^2.1.1"
PySimpleGUI = "^4.60.5"
avro-validator = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
